+++ src/Data/TimeSeries/Series.hs
@@ -12,13 +12,15 @@ module Data.TimeSeries.Series
     , Series
     , Timestamp
     , Value
-    , valueAt
+    , max
+    , min
     , series
     , size
     , slice
+    , valueAt
     ) where
 
-import Prelude
+import Prelude hiding (max, min)
 import Data.Word (Word64)
 
 
@@ -70,3 +72,24 @@ valueAt (Series xs) ts = safeHead [y | DP x y <- xs, x == ts]
 slice :: Series -> Timestamp -> Timestamp -> Series
 slice (Series xs) start end = Series [DP x y | DP x y <- xs, x >= start && x <= end]
 
+
+-- | Return maximum value in the series
+-- Complexity O(n)
+--
+-- >max (Series [DP 1 41.3, DP 2 52.22, DP 3 3.0]) == 52.22
+-- >max (Series []) == 0
+--
+max :: Series -> Value
+max (Series xs) = foldl (\d (DP _ y) -> max' d y) 0.0 xs
+    where max' a b = if a > b then a else b
+
+-- | Return maximum value in the series
+-- Complexity O(n)
+--
+-- >min (Series [DP 1 41.3, DP 2 52.22, DP 3 3.0]) == 3.0
+-- >min (Series []) == 0
+--
+min :: Series -> Value
+min (Series xs) = foldl (\d (DP _ y) -> min' d y) init xs
+    where min' a b = if a < b then a else b
+          init = fromIntegral (maxBound :: Int)