+++ test-src/Data/TimeSeries/CSVSpec.hs
@@ -1,0 +1,17 @@
+module Data.TimeSeries.CSVSpec (spec) where
+
+import Test.Hspec
+import Data.ByteString.Time (parseISODateTime)
+
+import qualified Data.TimeSeries as TS
+import qualified Data.TimeSeries.CSVReader as CSV
+
+
+spec :: Spec
+spec = do
+
+  describe "Reader" $ do
+
+    it "co2 test data" $ do
+        xs <- CSV.load parseISODateTime "testdata/co2.csv"
+        TS.size xs `shouldBe` 192