+++ CHANGES.md
@@ -1,6 +1,9 @@
 # Version 0.2 (2016-12-??)
 
-  * Implemented slice (Fixed #3)
+  * Implemented functions
+    * slice
+    * max
+    * min
 
 # Version 0.1 (2016-12-10)