+++ CHANGES.md
@@ -1,9 +1,10 @@
-# Version 0.2 (2016-12-??)
+# Version 0.2 (2016-12-13)
 
   * Implemented functions
     * slice
     * max
     * min
+  * Implemented CSV reader
 
 # Version 0.1 (2016-12-10)